refactor(base): split ImageText out of text/overlay.py (0.31.3)
Move the PIL rendering primitive (ImageText + TextAlign + AnchorPoint
+ its module-level constants and type aliases) into a new module
base/text/image_text.py. base/text/overlay.py shrinks from ~1090 LOC
to ~160 LOC and keeps only TranscriptionOverlay, which now imports
ImageText from the sibling module.

Re-exports through videopython.base.text and videopython.base keep
existing public imports working unchanged.

One internal test (test_text_overlay.py) was importing from the deep
path videopython.base.text.overlay; switch it to .image_text.
