+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "videopython"
-version = "0.31.2"
+version = "0.31.3"
 description = "Minimal video generation and processing library."
 authors = [
     { name = "Bartosz Wójtowicz", email = "bartoszwojtowicz@outlook.com" },
