+++ src/tests/base/test_text_overlay.py
@@ -4,7 +4,7 @@ from PIL import Image
 
 from tests.test_config import TEST_FONT_PATH
 from videopython.base.effects import FullImageOverlay
-from videopython.base.text.overlay import AnchorPoint, ImageText, TextAlign
+from videopython.base.text.image_text import AnchorPoint, ImageText, TextAlign
 
 
 def test_text_is_rendered_correctly():
